feat: implement a MAPS base dictionary item And rename a file that is incorrect
